he prbado todos los controllers, la semilla y la actualización

+++ src/api/controllers/places.js
@@ -1,24 +1,25 @@
 const { deleteFile } = require("../../utils/deleteFile");
 const Place = require("../models/places");
-const cloudinary = require("cloudinary");
+const mongoose = require("mongoose");
+
 
 const getPlaces = async (req, res, next) => {
     try {
         const places = await Place.find();
-        //preguntar como práctica el meter si no existe places para especificar más el error
+
         if(!places) {
-            return res.status(404).json("no hay sitios")
+            return res.status(404).json("No se encuentran estos sitios")
         }
         return res.status(200).json(places);
       
     } catch (error) {
-        return res.status(400).json({error: "error buscando sitios"})
+        return res.status(500).json({error:"Error buscando sitios", details: error.message})
     }
 }
 
 const getPlacesById = async (req, res, next) => {
     try {
-        const {id} = req.params;//preguntar por la mejor práctica para esto
+        const {id} = req.params;
         const place = await Place.findById(id);
 
         if (!place) {
@@ -26,19 +27,21 @@ const getPlacesById = async (req, res, next) => {
         }
         return res.status(200).json(place);
     } catch (error) {
-        return res.status(400).json({error: error.message})
+        console.log(error);
+        return res.status(400).json({error:"Error obteniendo los sitios" ,details: error.message})
     }
 }
 
 const postPlaces = async(req, res, next) =>{
     try {
-         if (!req.file) {
-        return res.status(400).json({ error: "No se envió ninguna imagen" });
-         }
         const newPlace = new Place({
             ...req.body, 
             img:{url: req.file.path}
         });
+
+        if (!req.file) {
+        return res.status(400).json({ error: "No se envió ninguna imagen" });
+         }
         
         /*if (req.user.rol === "admin") {
             newPlace.verified = true;
@@ -49,10 +52,10 @@ const postPlaces = async(req, res, next) =>{
         const placeSaved = await newPlace.save();
         return res.status(201).json(placeSaved);
     } catch (error) {
-        return res.status(400).json({error: "Error publicando un nuevo lugar"
-        })
+        console.error(error);
+        return res.status(500).json({ error: "Error publicando un nuevo lugar", details: error.message });
+        }
     }
-}
 
 const putPlaces = async (req, res, next) => {
     try {
@@ -71,24 +74,37 @@ const putPlaces = async (req, res, next) => {
     return res.status(200).json(updatedPlace);
 
     } catch (error) {
-        return res.status(400).json({error: "Error actualizando el lugar"})
+        console.log(error);
+        return res.status(500).json({error: "Error actualizando el lugar", details: error.message})
     }
 }
 
 const deletePlace = async(req, res, next) => {
     try {
         const {id} = req.params;
+        console.log(req.params.id);
+        //preguntar por la validación que hace Santi
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(400).json({ message: "ID no válido" });}
         const placesDeleted = await Place.findByIdAndDelete(id);
 
-        deleteFile(placesDeleted.img);
-
         if (!placesDeleted) {
-            return res.status(404).json("No se encuentra este lugar para borrarlo")
-        }
+            return res.status(404).json("No se encuentra este lugar para borrarlo");
+        };
 
+         if (placesDeleted.img?.public_id) {
+      try {
+        const result = await cloudinary.uploader.destroy(placesDeleted.img.public_id);
+        console.log("Imagen eliminada:", result);
+      } catch (error) {
+        console.log(error)
+        return res.status(500).json({error: "Error eliminando imagen en Cloudinary:", details: error.message});
+      }
+    }
         return res.status(200).json(placesDeleted)
     } catch (error) {
-        return res.status(400).json("Error eliminando el lugar")
+        console.log(error)
+        return res.status(500).json({error: "Error eliminando el lugar", details: error.message});
     }
 }
